+++ src/main/java/monopoly/model/Player.java
@@ -4,10 +4,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -22,7 +20,7 @@ public class Player implements StreetOwner, CardOwner {
 	private String id;
 	private Game game;
 	private StringProperty name = new SimpleStringProperty();
-	private DoubleProperty balance = new SimpleDoubleProperty();
+	private IntegerProperty balance = new SimpleIntegerProperty();
 	private ObservableList<Street> streets = FXCollections.observableArrayList();
 	private ObservableList<Card> cards = FXCollections.observableArrayList();
 	private IntegerProperty jailCount = new SimpleIntegerProperty();
@@ -58,15 +56,15 @@ public class Player implements StreetOwner, CardOwner {
 		this.name.set(name);
 	}
 
-	public DoubleProperty getBalance() {
+	public IntegerProperty getBalance() {
 		return balance;
 	}
 
-	public void pay(double amount) {
+	public void pay(int amount) {
 		balance.set(balance.get() - amount);
 	}
 
-	public void addMoney(double amount) {
+	public void addMoney(int amount) {
 		balance.set(balance.get() + amount);
 	}
 
@@ -134,7 +132,7 @@ public class Player implements StreetOwner, CardOwner {
 		cards.remove(card);
 	}
 
-	public void setBalance(double balance) {
+	public void setBalance(int balance) {
 		this.balance.set(balance);
 	}
